+++ tests/testthat/test-warning.R
@@ -85,7 +85,7 @@ test_that("warning raised if outputs exceed threshold", {
 
   output <- a_output
   output$indicators$mean[
-    output$indicators$indicator == "prevalence"][1] <- 0.5
+    output$indicators$indicator == "prevalence"][1] <- 0.6
   output$indicators$mean[
     output$indicators$indicator == "art_coverage"][1] <- 1.2
   mock_output_package <- mockery::mock(output)
@@ -100,7 +100,7 @@ test_that("warning raised if outputs exceed threshold", {
   expect_length(out$warnings, 2)
   expect_equal(
     out$warnings[[1]]$text,
-    "HIV prevalence is higher than 40% for: March 2016, Northern, Both, 0-4")
+    "HIV prevalence is higher than 50% for: March 2016, Northern, Both, 0-4")
   expect_equal(
     out$warnings[[2]]$text,
     "ART coverage is higher than 100% for: March 2016, Northern, Both, 0-4")